refactor(frontend): 抽取 useVideos composable + 命名路由
- AdminLayout router-link 改用命名路由
- 新增 useVideos composable，统一 VideoItem 类型和数据加载
- 新增 isTouhou()，从 touhouStatusOptions 推导，不硬编码状态值
- DashboardHome 和 VideoManage 共用 useVideos，去掉重复定义
